All Api routes secured and authenticated for particular user

+++ controller/contactController.js
@@ -2,15 +2,15 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel")
 //@desc Get all contacts
 //@route GET /api/contacts
-//@acess public
+//@acess private
 const getContacts = asyncHandler(async (req,res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id : req.user.id});
     res.status(200).json(contacts);
 });
 
 //@desc Create new contacts
 //@route POST /api/contacts
-//@acess public
+//@acess private
 const createContact = asyncHandler(async (req,res) => {
     // console.log("the req body is :",req.body);
     const {email, name, phone} = req.body;
@@ -21,14 +21,15 @@ const createContact = asyncHandler(async (req,res) => {
     const contact = await Contact.create({
         name,
         email,
-        phone
-    })
+        phone,
+        user_id : req.user.id,
+    });
     res.status(201).json(contact);
 });
 
 //@desc Get contact
 //@route GET /api/contacts/:id
-//@acess public
+//@acess private
 const getContactwithID = asyncHandler(async (req,res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -40,14 +41,17 @@ const getContactwithID = asyncHandler(async (req,res) => {
 
 //@desc Update contact
 //@route PUT /api/contacts/:id
-//@acess public
+//@acess private
 const updateContact = asyncHandler(async (req,res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not Found");
     }
-
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(404);
+        throw new Error("User don't have the permission to update other user contacts")
+    }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -58,12 +62,16 @@ const updateContact = asyncHandler(async (req,res) => {
 
 //@desc Delete contact
 //@route DELETE /api/contacts/:id
-//@acess public
+//@acess private
 const deleteContact = asyncHandler(async (req,res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact not Found");
+    }
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(404);
+        throw new Error("User don't have the permission to delete other user contacts")
     }
     await Contact.deleteOne({_id : req.params.id});
     res.status(200).json(contact);
